send image work

+++ app/src/main/java/edu/northeastern/team1/Message.java
@@ -1,17 +1,17 @@
 package edu.northeastern.team1;
 
 public class Message {
-    private Integer mid;
+    private Long mid;
     private String sender;
     private String image;
 
-    public Message(Integer mid, String sender, String image) {
+    public Message(Long mid, String sender, String image) {
         this.mid = mid;
         this.sender = sender;
         this.image = image;
     }
 
-    public Integer getMid() {
+    public Long getMid() {
         return mid;
     }
